requirements.txt
boto3==1.34.29
python-dotenv==1.0.0
click==8.1.7 